ecommerce/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.20.46
botocore==1.23.46
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==9.0.0
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==5.3.0
mysqlclient==2.1.1
django-filter==22.1
django-mysql==4.8.0
uwsgi>=2.0.19<2.1
